Move `Metadata` and `Data` classes to `files` module
Avoids import of `anyio` without actually using the `SQLiteStore` component
